LIBDRUM-659. Added "drum" custom theme, based on "dspace" theme
Created a "drum" theme for customizing DSpace front-end for UMD DRUM
website. The expectation is that there will also be a separate them
for the "MD-SOAR" website at some point in the future.

Since we can "extend" themes, and the "dspace" theme is the default, it
seemed to make the most sense to use the "dspace" theme as the "base"
theme from which to create the "drum" theme.

Copied the "header" and "navbar" components from the "dspace" theme,
and modified to display the UMD and DRUM logos. The small "UMD" logo
was used, instead of the "UMD Libraries" logo, because the greater
width of the "UMD Libraries" logo was causing layout issues with the
navigation bar (specifically, that at narrower widths, the "Login" link
would go past the left edge of the browser window).

https://issues.umd.edu/browse/LIBDRUM-659