LIBDRUM-659. Modified "config.server.ts" to read "config.yml" file
According to the DSpace 7 documentation at

https://wiki.lyrasis.org/display/DSDOC7x/User+Interface+Configuration,

values in the "config/config.yml" file should be overridden by values in
the dev/prod/test environment configuration
 "config/config[dev/prod/test].yml" file, if such a file exists. This
code change fixes an issue in which the "config/config.yml" file would
not be loaded, if a "config/config[dev/prod/test].yml" exists.

It does this by adding a "getBaseConfigPath", which looks for the
"config/config.yml" (or "config/config.yaml") file, and loads it. The
"getLocalConfigPath" method has been modified to only return the
environment-specific "config/config[dev/prod/test].yml", if it exists.

https://issues.umd.edu/browse/LIBDRUM-659