LIBDRUM-660. Added "UM Directory Information" section to EPerson form
On the EPerson edit form, added a "UM Directory Information" section,
with information about the EPerson retrieved from LDAP (if the eperson
is known to LDAP).

Added an "ldap" link to the EPerson model, and added an LdapDataService
to handle retrieving information from the link.

Added a "stub" LdapDataService to the tests, so that the existing
EPerson tests would pass.

https://issues.umd.edu/browse/LIBDRUM-660