LIBDRUM-671. Fix to handle creation of new ETD department
Modified "etdunit-form.component.html" to suppress display of the
Collection membership and search form when a new ETD Department is
being created.

Modified "etdunit-form.component.ts" to fix a spurious request to the
DSpace backend for an ETD department named "newEtdUnit" or "undefined"
when creating a new ETD department. Note: The "Units" functionality
exhibits the same behavior, so we might want to think about making
a similar change to that code as well.

https://umd-dit.atlassian.net/browse/LIBDRUM-671