LIBDRUM-678. Added "restricted-access" module and page
Added a "restricted-access" module that shows the "Restricted Access"
page when a bitstream is embargoed, or otherwise not downloadable by
the user (such as due to a "Campus" IP restriction, and the user is
off-campus).

Displays a "(RESTRICTED ACCESS)" text marker below the download link
whenever the bitstream has an active embargo (shown for all users, even
if the user can actually download the bitstream).

The "Restricted Access" page show differing text based on whether the
file has a time-limited embargo, a forever embargo, the user is logged
in, or if there is some other (non-embargo) restriction.

https://umd-dit.atlassian.net/browse/LIBDRUM-678