LIBDRUM-678. Modified restricted access routing to retrieve bitstream
Modified the "RestrictedAccessPageRoutingModule" to use the
BitstreamPageResolver to retrieve the bitstream so that information
about the bitstream (such as embargo date or filename) can be used in
the restricted access page.

Modified the "BitstreamDownloadPageComponent to include the bitstream
UUID in the navigation to the restricted access page, so that it can
be used to retrieve the bitstream.

Modified the "RestrictedAccessComponent" to retrieve the appropriate
messages and display, based on the user's login status, and the embargo
restrictions.

https://umd-dit.atlassian.net/browse/LIBDRUM-678