LIBDRUM-715. Added "JsonLdWebsiteComponent" to support "WebSite" JSON-LD
Added “JsonLdWebsiteComponent” class which generates the JSON-LD object
and uses the “JsonLdService” to insert it into the web page.

Modified the “JsonLdWebsiteComponent” component to the “DECLARATIONS”
constant in “src/themes/drum/eager-theme.module.ts”. The “eager” as
opposed to “lazy” module configuration was used because this component
will be used on every page.

Added the “ds-json-ld-website” component HTML tag to the
“src/themes/drum/app/header/header.component.html” file, so the component
would be activated on every web page.

While it is technically possible to retrieve the “url” property from the
“document.location.origin” property of the HTML document being displayed
which is available to the Angular code), the DSpace 6 code uses the
“dspace.baseUrl” property, and this seemed like a safer approach.

In DSpace 6, the “name” property is hard-coded in to the
“page-structure.xsl” document, and this did not seem optimal for the
Angular code. So instead opted to use the “dspace.name” property.

https://umd-dit.atlassian.net/browse/LIBDRUM-715