LIBDRUM-734. Switch order of log in options on Log In menu
Modified the "AuthInterceptor" class to place the "CAS" authentication
method as the first entry in the sorted array, instead of "password".
This ensures that that the "CAS" option will appear first in the
"LoginComponent" drop-down list.

Modified button labels in the "en.json5" file to match the requests
in the issue.

https://umd-dit.atlassian.net/browse/LIBDRUM-734