LIBDRUM-760. Refactor I18n customizations into global "en.json5 file"
Moved all I18n customizations into the DSpace "global" I18n file
at "src/assets/i18n/en.json5", placing them into a "UMD Customization"
section at the bottom of the file.

This takes advantage of the behavior that when multiple instance of the
same key are in the file, the "last" instance of the key is used.

Deleted the "theme" I18n file in "src/themes/drum/assets/i18n/en.json5"
after verifying that all changes were present in the "UMD Customization"
section of "src/assets/i18n/en.json5".

It was decided to modify the global "src/assets/i18n/en.json5" file
directly, instead of using a "theme" file, which would require running
the "merge-i18n" script whenever it was changed (which in turn would
modify the "global" I18n file). It just seemed simpler to edit the
"global" I18n file directly (in a "customization" block at the end of
the file), instead of relying on developers to remember to run the
"merge-i18n" script when needed.

https://umd-dit.atlassian.net/browse/LIBDRUM-760