LIBDRUM-745. Updated DRUM footer
Modified the “src/themes/drum/app/footer/footer.component.html” file,
keeping the “top” and “bottom” footer structure, but removing the
DSpace copyright and Lyrasis link.

Reordered the row of links in the bottom footer to
“Send Feedback | Privacy policy | Web Accessibility | Give Now | Cookie settings”
and removed the “text-white” class, so that the text color would match
the top footer.

Modified the “src/themes/drum/app/footer/footer.component.scss” file,
changing the “background-color” for the “footer” class
(the bottom footer) to use the same SCSS variable as the top footer,
so that the background colors would match.

Considered merging the “top” and “bottom” footer sections together to be
a single footer, but DSpace seems geared to have both sections, and the
CSS styling is different enough that it did not seem to be worth the
effort, especially if we want to make changes in the future (or keep up
with any DSpace upstream changes).

https://umd-dit.atlassian.net/browse/LIBDRUM-745