LIBDRUM-746. Removed "preserve-line-breaks" on Item page metadata fields
Modified the metadata field display on the "Item" page by removing
the "preserve-line-breaks" CSS class, so that the multi-line "Abstract"
and "Notes" fields are displayed as single blocks of text, ignoring
any line breaks in the text. This is to make the display more similar
to that of DSpace 6, which had the same behavior.

Without this change, the "Abstract" and "Notes" fields of many items
displayed oddly, because the line breaks were not compatible with the
available display width, leading to many short line fragments.

https://umd-dit.atlassian.net/browse/LIBDRUM-746