LIBDRUM-748. Wufoo feedback form implementation
Wufoo form redirect implemented using an Angular “resolver” component
(src/app/wufoo/wufoo-feedback-resolver.ts).

The WufooFeedbackResolver is run as part of the route navigation process
when the “Send Feedback” link is clicked by the user. It retrieves the
page on which the user clicked the link, and provides it to the
"wufoo-feedback" endpoint on the back-end, which returns a URL
pr-populates with the form default values as query parameters.

The “src/app/info/info-routing.module.ts” routing module was modified,
adding the “WufooFeedbackResolver” to the resolvers for the
“/info/feedback” link, so that the redirect is triggered.

Modified the "Send Feedback" link in
"src/themes/drum/app/footer/footer.component.html" so that the browser
URL would not be modified as part of the navigation. This enables the
user to navigate back to the original page using the browser's "Back"
button.

The “src/app/wufoo/wufoo-feedback-response.model.ts” is a simple model
for processing the redirect URL response from the back-end.

As the “feedback-component” is themed, copied the
“src/app/info/feedback” directory into
“src/themes/drum/app/info/feedback”, and made the following
modifications:

* Removed all the logic from the “FeedbackFormComponent” class, as it is
  no longer needed.
* Deleted the “feedback-form.component.spec.ts” file, as the tests were
  no longer relevant
* Replaced the HTML in “feedback-form.component.html” with simple
  placeholder text. This text will is shown (briefly) when the redirect
  occurs, and will be shown if the “wufoo-feedback” request fails for some reason.

https://umd-dit.atlassian.net/browse/LIBDRUM-748