LIBDRUM-772. Replaced university/DRUM logos with new combined logo
Replaced the separate university and DRUM logos with a single combined
logo that links to the DRUM home page.

The "drum-logo.png" DRUM logo PNG image was replaced with an SVG image
"drum-logo.svg", and the "umd_informal_seal.gif" file was removed as
no longer necessary.

The SVG file from the issue was used, instead of the PNG image, because
it is significantly smaller. However, since SVG images are not correctly
sized by the "max-height" attribute (and would not be shown, as being
of zero height), the CSS files were modified to add a "height" attribute
to the "navbar.component.scss" file.

Also, because the DSpace default height of 50px was too small, changed
the "ds-header-logo-height" in "_theme_css_variable_overrides.scss" to
"70px".

The HTML elements for the logos in the "navbar.component.html" and
"header.component.html" were replaced with a single logo element.

Note: Is it unclear whether/how the image in "header.component.html"
is displayed, but it is present in the "dspace" theme version of the
file, and so was modified.

https://umd-dit.atlassian.net/browse/LIBDRUM-772