LIBDRUM-785. Implement two-column layout on DRUM home page
Modified the home page to show "List of Communities" and
"Recent Submissions" in a two-column layout.

The 2-column implementation required making changes to the stock
DSpace “recent-item-list” component, which is not themeable. Therefore,
the stock DSpace file
“src/app/home-page/recent-item-list/recent-item-list.component.html”
was customized in-place. The change in the file was adding Bootstrap
tags to suppress the margin top and dividing line in the “small” layouts
and above, so that the  “Recent Submissions” title in the second column
would line up with the “List of Communities” title  in the first column.

The “src/themes/drum/app/home-page/home-page.component.html” was also
modified to place the “List of Communities” and “Recent Submissions”
nto a two-column layout, using Bootstrap grid functionality.

https://umd-dit.atlassian.net/browse/LIBDRUM-785