LIBDRUM-812. Updates to fix Angular compile errors for themed components
Added additional modules to "src/themes/drum/eager-theme.module.ts" to
fix Angular compile errors. The errors appear to be related to replacing
some of the components in the "drum" theme HTML with their themed
component counterparts.

an error in the browser when using the "ds-themed-feedback-form"
component, instead of "ds-feedback-form". Since we don't use the
DSpace-provided feedback form (instead, redirecting to an external
Wufoo form), this should not be an issue.

https://umd-dit.atlassian.net/browse/LIBDRUM-812