LIBDRUM-827. Keep bitstream download count separate from file hyperlink
In the DSpace 7.6 upgrade, this file was modified to match the stock
DSpace file in which the "ngFor" iterator was part of the
"ds-themed-file-download-link" component, and the
"ds-bitstream-download-counter" component was part of the
"ds-themed-file-download-link" content. This caused an issue in which
the bitstream download count was included as part of the file hyperlink.

Fixed the issue by moving the "ngFor" iterator into a separate
"ngContainer", and moving the "ds-bitstream-download-counter" component
out of the content of the "ds-themed-file-download-link" component.

https://umd-dit.atlassian.net/browse/LIBDRUM-827