LIBDRUM-842. Restore DRUM logo to navigation bar
Incomplete merging of the DSpace 7.6.1 changes to the navigation bar
HTML and SCSS was causing the DRUM logo to not appear in wider screens.

Migrated DSpace 7.6.1 changes in the "src/themes/dspace" theme into the
"src/themes/drum" theme.

Also updated the "ds-header-navbar-border-bottom-color" SCSS variable in
the "_theme_css_variable_overrides.scss" file to use the DRUM-preferred
color.

https://umd-dit.atlassian.net/browse/LIBDRUM-842