LIBDRUM-843. Replace DSpace logo on the /login page with DRUM logo
Added the "login-page" component to the "drum" theme by copying the
files from the "custom" theme. Modified the files to replace the
DSpace logo on the /login page with the DRUM logo.

Added "--ds-login-logo-height/--ds-login-logo-width" to the
"src/themes/drum/styles/_theme_css_variable_overrides.scss" file to
set the correct width/height for the DRUM logo on the page.

Updated the "docs/DrumAngularCustomization.md" document.

https://umd-dit.atlassian.net/browse/LIBDRUM-843