LIBDRUM-838. Added "Submit item to DRUM" menu entry to navigation bar
Added a "Submit item to DRUM" menu entry to the navigation bar that
displays for logged-in users that have "submit" permission to at least
one collection.

This menu entry is essentially a duplicate of the "New | Item" menu
entry in the administrative sidebar.

The "index" is set to "3" so that it will show at the end of the
navigation bar entries (the index "2" is used by the "Statistics" menu
entry).

https://umd-dit.atlassian.net/browse/LIBDRUM-839