LIBDRUM-797. Fix "Show More" button for subcommunities
In "community-list.component.ts", fixed the "Show More" button when
displayed in subcommunities such as the
“MARAC Mid-Atlantic Regional Archives Conference” community
by replacing the use of the "getMaxAllowedCommunities" method with the
"nextPage" method. The "nextPage" method was modified to retrieve
up to "MAX_COMCOLS_PER_PAGE" entries, instead of the DSpace default of
5 entries. The "getMaxAllowedCommunities" method was no longer needed,
and so was removed

Replaced the use of "getMaxAllowedCommunities" in the
"community-list.component.html" file with "getNextPage". This restores
the HTML file to being identical with the stock DSpace version, but
retaining the HTML file because this really represents the HTML
for the UMD custom "ds-cg-community-list" component, not the stock
DSpace "ds-community-list", so the similarity is coincidental, not
duplicative.

https://umd-dit.atlassian.net/browse/LIBDRUM-797