LIBDRUM-857. Replace home page background image with tiled image
Replaced the 4.27 megabyte background image on the home page with a
tiled image that is 43.6 kilobytes (nearly 1/100th the size).

Because of the use of tiling, the subtle northwest-to-southwest gradient
that subtly fades the foreground design that existed in the larger
background is not replicated. This gradient was not considered essential
for this issue.

Another difference is that the foreground design in the image does not
resize as the browser window width changes. The change in size seen in
the original image was also not deemed essential for this image.

Getting the tiled image to actually tile required removing the
“picture/img” HTML tag hierarchy, which included the image filename in
the “src/themes/drum/app/home-page/home-news/home-news.component.html”
(originally derived from the “dspace” theme), and replacing it with CSS
settings in the
“src/themes/drum/app/home-page/home-news/home-news.component.scss” file
(the image filename is specified by a “background-image” CSS element.

https://umd-dit.atlassian.net/browse/LIBDRUM-857