LIBDRUM-857. Fix background image file location for Angular production
When running on the https://drum-test.lib.umd.edu/ the
"drum-background-image-tile.png" was requiring two requests to find.
The first to the "drum-background-image-tile.png" file was returning
a 404 error, followed by a second request to an Angular-derived
version of the file with a "cache-busting" extension that succeeded.

Based on guidance in https://stackoverflow.com/a/48890991, changed
the "background-image" URL to be an absolute, instead of relative
path, which fixes the issue seen in production, while still enabling
the image to be retrieved in development.

https://umd-dit.atlassian.net/browse/LIBDRUM-857