LIBDRUM-879. Added "NativeWindowService" for sidebar class constructors
In LIBDRUM-872, the sidebar classes were modified to include a
"NativeWindowService" parameter in their constructors, so that the
browser and operating system could be determined. It was overlooked in
that issue that the unit tests also needed to be updated.

Modified the unit tests to include a mock NativeWindowService instance
as the "NativeWindowService" parameter provided to the class
constructors.

https://umd-dit.atlassian.net/browse/LIBDRUM-879