LIBDRUM-879. Place the z-index of UMD banner below modal dialogs
Placed the z-index of UMD banner below that of the modal dialogs.
This enables modal dialogs to overlap the UMD banner, if necessary,
instead of being obscured by it.

This also ensures that Cypress end-to-end tests that check accessibility
will pass successfully.

Need the full "body div.umdheader-wrap" selector to override the
"umdheader-wrap" selector provided with the UMD banner.

Decided to set the z-index to ($zindex-modal - 100), as that would put
it beneath any of the other Bootstrap-specified z-indexes
(see https://getbootstrap.com/docs/5.0/layout/z-index/).

https://umd-dit.atlassian.net/browse/LIBDRUM-879