LIBDRUM-873. Enable text wrapping in navigation bar entries
This commit consists of the changes from dspace-angular Pull Request
3513, incorporated into the appropriate "drum" and "dspace" theme files.

The changes in Pull Request 3513 enable the entries in the navigation
bar to perform text wrapping at smaller browser widths (similar to the
behavior in DSpace 7.6.1) so that the right-most navigation bar entries
are available at nearly all browser widths.

There are still some browser widths at which the right-most entry is
fully obscured, but this same behavior was also seen in DSpace 7.6.1.

https://umd-dit.atlassian.net/browse/LIBDRUM-873