LIBDRUM-870. Suppress display of license bundle on full item page
DRUM provides the ability for users to upload their own licenses which
may contain user-specific information such as a user’s email address.

In DSpace 6, the license file was not downloadable, and we wish to
continue that policy in DSpace 7, to avoid potentially exposing
user-specific information.

From discussions with Terry Owen, removing the “License bundle” section
entirely for all users (including Administrators) is acceptable.
Administrators will still have access to the files via the “Bitstreams”
tab on the “Edit Item” page (with access to that page controlled by
DSpace’s access control mechanisms).

DSpace does not seem to provide bitstream-specific access controls to
files in the “License bundle”, so if a user were to somehow obtain the
URL to the license file, they would still be able to download it. We are
not particularly worried about this scenario, as the URLs use opaque
UUID-like identifiers that are likely not guessable, so attempting to
prevent this is not part of this issue.

Modified the
“src/app/item-page/full/field-components/file-section/full-file-section.component.ts”
file directly to simply suppress the retrieval of the license bundle
from the back-end, instead of moving the files into the “drum” theme and
modifying there, because it seemed more straight-forward, and it would
likely simplify future upgrades.

Also modified the
“src/app/item-page/full/field-components/file-section/full-file-section.component.spec.ts”
test to account for the license bundle files no longer being downloaded.

Updated the “docs/DrumAngularCustomization.md” document.

https://umd-dit.atlassian.net/browse/LIBDRUM-870