LIBDRUM-890. Suppress "Researcher Profile" panel on "Update Profile"
This code change to suppress the "Researcher Profile" panel on
the "Update Profile" page is taken from commit "ded0079" of
dspace-angular pull request 3585.

Customization markers have been added to indicate that the changes
came from a stock DSpace pull request, and can be removed when DRUM
is upgraded to a version incorporating the changes.

https://umd-dit.atlassian.net/browse/LIBDRUM-890