LIBDRUM-892. Replaced "ds-cg-community-list" with "ds-community-list"
In DSpace 7.x, a UMD-add "ds-cg-community-list" selector was used for
the "CommunityListComponent" TypeScript class, due to a selector
conflict with the DSpace-provided class.

In this change, a DSpace 8.0 Angular lint error
("dspace-angular-ts/themed-component-selectors") instigated a
refactoring where the DRUM "CommunityListComponent" theme class now
uses the "ds-themed-base-community-list" selector, instead of
the "ds-cg-community-list" selector, and an override of the stock
"ThemedCommunityListComponent" class was added to ensure that the
correct DRUM version of the component is used. The override of the
"ThemedCommunityListComponent" was necessary, because of the additional
"@Input" fields needed by the DRUM "CommunityListComponent" class.

Updated the imports and HTML tags as needed.

https://umd-dit.atlassian.net/browse/LIBDRUM-892