LIBDRUM-892. Skip "community-create-cy.ts" Cypress e2e test
This test was failing with the following:

```
cypressError: `cy.type()` failed because it targeted a disabled element.
```

This is because the community form has been modified by UMD to include
a required "Community Group" field, that cannot be populated from the
stock DSpace database used by the e2e tests when run by GitHub actions.

Skipping the test so that GitHub Actions is can still succeed.

https://umd-dit.atlassian.net/browse/LIBDRUM-892