LIBDRUM-892. Update DRUM community-list.component.html for DSpace 8.0
Incorporated changes in the stock DSpace 8.0
"community-list.component.html" file with the matching file in
the DRUM theme.

There are no UMD customizations in the DRUM file, but keeping it for the
moment, because there it seems more consistent to keep it, than to
modified the component class to point to the stock DSpace 8.0 file.

Also fixed a minor typo in the "community-list.component.ts" file.

https://umd-dit.atlassian.net/browse/LIBDRUM-892