LIBDRUM-902. Fixed "Delete ETD Department" button display on create form
Modified the check for displaying the "Delete ETD Department" button to
take into account whether the department exists.

Also simplified the "translateService" call by removing an
"observableCombineLatest" operator that only had one argument, and
fixed a deprecation warning on a second "observerCombineLatest" operator
by converting the arguments to an array.

https://umd-dit.atlassian.net/browse/LIBDRUM-902